Moved IsProxyOf to Basebuildings
The ported code compared the wrong entities. Now the code is directly implemented in basebuildings and fixes the old bug again, that building a mega-command-centre removes dependencies of the comman-centre.

Also why not allow CCs to be demolished? Sometimes you need the place (also its easier to test :stuck_out_tongue_closed_eyes: